fix some compile warnings for gcc 4.1

+++ src/resmom/linux/mom_mach.c
@@ -1215,7 +1215,7 @@ int mom_set_limits(
         if (LOGLEVEL >= 2)
           {
           sprintf(log_buffer,"setting cpu time limit to %ld for job %s",
-            reslim.rlim_cur,
+            (long int)reslim.rlim_cur,
             pjob->ji_qs.ji_jobid);
 
           log_record(PBSEVENT_SYSTEM,0,id,log_buffer);
@@ -1257,7 +1257,7 @@ int mom_set_limits(
           if (LOGLEVEL >= 0)
             {
             sprintf(log_buffer,"cannot set file limit to %ld for job %s (value too large)",
-              reslim.rlim_cur,
+              (long int)reslim.rlim_cur,
               pjob->ji_qs.ji_jobid);
 
             log_err(-1,id,log_buffer);
@@ -1273,7 +1273,7 @@ int mom_set_limits(
         if (setrlimit(RLIMIT_FSIZE,&reslim) < 0)
           {
           sprintf(log_buffer,"cannot set file limit to %ld for job %s (setrlimit failed - check default user limits)",
-            reslim.rlim_max,
+            (long int)reslim.rlim_max,
             pjob->ji_qs.ji_jobid);
 
           log_err(errno,id,log_buffer);
@@ -1357,7 +1357,7 @@ int mom_set_limits(
         if (setrlimit(RLIMIT_DATA,&reslim) < 0)
           {
          sprintf(log_buffer,"cannot set data limit to %ld for job %s (setrlimit failed w/errno=%d - check default user limits)",
-            reslim.rlim_max,
+            (long int)reslim.rlim_max,
             pjob->ji_qs.ji_jobid,
             errno);
 
@@ -1367,7 +1367,7 @@ int mom_set_limits(
         if (setrlimit(RLIMIT_RSS,&reslim) < 0)
           {
          sprintf(log_buffer,"cannot set RSS limit to %ld for job %s (setrlimit failed w/errno=%d - check default user limits)",
-            reslim.rlim_max,
+            (long int)reslim.rlim_max,
             pjob->ji_qs.ji_jobid,
             errno);
 
@@ -1380,7 +1380,7 @@ int mom_set_limits(
         if (setrlimit(RLIMIT_STACK,&reslim) < 0)
           {
           sprintf(log_buffer,"cannot set stack limit to %ld for job %s (setrlimit failed w/errno=%d - check default user limits)",
-            reslim.rlim_max,
+            (long int)reslim.rlim_max,
             pjob->ji_qs.ji_jobid,
             errno);
 
@@ -1392,7 +1392,7 @@ int mom_set_limits(
         if (setrlimit(RLIMIT_AS,&reslim) < 0)
           {
           sprintf(log_buffer,"cannot set AS limit to %ld for job %s (setrlimit failed w/errno=%d - check default user limits)",
-            reslim.rlim_max,
+            (long int)reslim.rlim_max,
             pjob->ji_qs.ji_jobid,
             errno);
